Use an in-tree plugin interface instead of a Go plugin proper
I might have been missing something, but the Go plugin infrastructure
seems impossible to work with when you have plugins with complex
dependencies.  Even if there was a way to make it work, it doesn't seem
worth my time to investigate how to do that.  Simply use an in-tree
"plugin" model.